third_party/blink/web_tests/html/select/select-typeahead-scroll-expected.txt

PASS select.open is true
PASS successfullyParsed is true

TEST COMPLETE


Test that selection box scrolls to where the focus jumps when pressing an alphanumeric key.

aardvark
blue
navy
indigo
azure
SUCCESS
